+++ saleor/graphql/order/mutations/order_note_add.py
@@ -10,7 +10,7 @@ from ...app.dataloaders import get_app_promise
 from ...core import ResolveInfo
 from ...core.context import SyncWebhookControlContext
 from ...core.doc_category import DOC_CATEGORY_ORDERS
-from ...core.types import BaseInputObjectType, Error, OrderError
+from ...core.types import Error
 from ...plugins.dataloaders import get_plugin_manager_promise
 from ..types import Order, OrderEvent
 from .order_note_common import OrderNoteCommon
@@ -68,28 +68,3 @@ class OrderNoteAdd(OrderNoteCommon):
             order=SyncWebhookControlContext(order),
             event=SyncWebhookControlContext(event),
         )
-
-
-class OrderAddNoteInput(BaseInputObjectType):
-    message = graphene.String(
-        description="Note message.",
-        name="message",
-        required=True,
-    )
-
-    class Meta:
-        doc_category = DOC_CATEGORY_ORDERS
-
-
-class OrderAddNote(OrderNoteAdd):
-    class Arguments(OrderNoteAdd.Arguments):
-        input = OrderAddNoteInput(
-            required=True, description="Fields required to create a note for the order."
-        )
-
-    class Meta:
-        description = "Adds note to the order."
-        doc_category = DOC_CATEGORY_ORDERS
-        permissions = (OrderPermissions.MANAGE_ORDERS,)
-        error_type_class = OrderError
-        error_type_field = "order_errors"